==> 2. Converting Monthly to Yearly.ipynb <==
# Libraries

import numpy as np
import pandas as pd

# Conversion

ozone_month = pd.read_excel("datasets/ozone_month.xlsx")
ozone_month = ozone_month.rename(columns={"Unnamed: 0": "Name"})
ozone_month = ozone_month.set_index("Name")

pm25_month = pd.read_excel("datasets/pm25_month.xlsx")
pm25_month = pm25_month.rename(columns={"Unnamed: 0": "Name"})
pm25_month = pm25_month.set_index("Name")

def convert_year(df, start_year, end_year):
    "Given data per month, average it out per year"
    num_years = end_year - start_year + 1
    indices = np.array(range(0, num_years))
    indices = indices * 12 # indices for the January for each year
    new_df = pd.DataFrame(index = df.index, columns = list(range(start_year, end_year + 1)))
    for i in range(0, num_years):
        indices_for_mean = list(range(indices[i], indices[i] + 12))
        new_df.iloc[:,i] = pd.DataFrame.mean(df.iloc[:,indices_for_mean], axis=1)
    return new_df

ozone_year = convert_year(ozone_month, 2005, 2018)
pm25_year = convert_year(pm25_month, 2005, 2018)

ozone_year.to_excel("datasets/ozone_year.xlsx")
pm25_year.to_excel("datasets/pm25_year.xlsx")
